+++ cd-extras/public/Aliases.ps1
@@ -8,6 +8,7 @@ Set-Alias cd- Undo-Location
 Set-Alias cd+ Redo-Location
 Set-Alias cd: Switch-LocationPart
 Set-Alias dirs Get-Stack
+Set-Alias dirsc Clear-Stack
 Set-Alias setocd Set-CdExtrasOption
 
 Set-Item Alias:cd Set-LocationEx