+++ cd-extras/public/Expand-Path.ps1
@@ -18,9 +18,6 @@ Limits search results to leaf items.
 .PARAMETER Directory
 Limits search results to container items.
 
-.ALIASES
-xpa
-
 .EXAMPLE
 # Expand a well-known Windows path by abbreviating each segment
 PS C:\> Expand-Path /win/sys/dr/et -Directory
@@ -64,7 +61,7 @@ function Expand-Path {
   $wildcardedPaths = if ($SearchPaths -and -not ($Candidate | IsRootedOrRelative)) {
     # always include the local path, regardless of whether it was passed
     # in the searchPaths parameter (this differs from the behaviour in bash)
-    @($wildcardedPath) + ($SearchPaths | Join-Path -ChildPath $wildcardedPath)
+    $wildcardedPath, ($SearchPaths | Join-Path -ChildPath $wildcardedPath)
   }
   else { $wildcardedPath }