+++ cd-extras/public/Undo-Location.ps1
@@ -9,9 +9,6 @@ The number of locations to undo.
 .PARAMETER NamePart
 Partial path name to choose from undo stack.
 
-.ALIASES
-cd-
-
 .EXAMPLE
 PS C:\Windows\System32> # Move backward to the previous location
 PS C:\Windows\System32> cd ..